+++ src/sys/mod.rs
@@ -70,6 +70,9 @@ pub mod statfs;
 
 pub mod statvfs;
 
+#[cfg(any(target_os = "android", target_os = "linux"))]
+pub mod sysinfo;
+
 pub mod termios;
 
 pub mod time;